+++ packages/antd/src/not-antd/index.ts
@@ -4,5 +4,5 @@ export { UIFragmentSpec, extractUIFragmentFromSpec } from './UIFragmentSpec';
 export { Anchor } from './Anchor';
 export { rootOrDetails, DetailProps } from './rootOrDetails';
 export { renderFallback } from './rendering';
-export { Redirect, redirect, redirectTo } from './Redirect';
+export { Redirect, redirect, redirectTo, redirectToNavRef } from './Redirect';
 export { NavRefRedirect } from './NavRefRedirect';