+++ packages/antd/src/routing/LinkAnt.tsx
@@ -1,5 +1,5 @@
 import { Location as MyLocation } from 'history';
-import { CoreLinkProps, UsesLocation, locationToUrl } from '@moxb/moxb';
+import { CoreLinkProps, UsesLocation, locationToUrl, NavRef, NavRefCall } from '@moxb/moxb';
 import { Button } from 'antd';
 import { ButtonProps } from 'antd/lib/button';
 import { inject, observer } from 'mobx-react';
@@ -8,6 +8,13 @@ import * as Anchor from '../not-antd/Anchor';
 
 export type LinkAntProps = Anchor.AnchorParams &
     CoreLinkProps & {
+        /**
+         * Should this link point to a NavRef?
+         *
+         * When given, this will override to, appendTokens, etc..
+         */
+        toRef?: NavRef<void | {}> | NavRefCall<any>;
+
         /**
          * How do you want to render this link? (Optional; defaults to "anchor")
          *
@@ -40,17 +47,34 @@ export class LinkAnt extends React.Component<LinkProps & UsesLocation> {
      * Calculate the location where this links should take us
      */
     protected getWantedLocation(): MyLocation {
-        const { position, to, argChanges, appendTokens, removeTokenCount } = this.props;
+        const { position, to, argChanges, appendTokens, removeTokenCount, toRef } = this.props;
         const locationManager = this.props.locationManager!;
-        const startLocation = appendTokens
-            ? locationManager.getNewLocationForAppendedPathTokens(appendTokens)
-            : removeTokenCount
-            ? locationManager.getNewLocationForRemovedPathTokens(removeTokenCount)
-            : locationManager.getNewLocationForPathAndQueryChanges(undefined, position, to, undefined);
-        return (argChanges || []).reduce(
-            (prevLocation: MyLocation, change) => change.arg.getModifiedLocation(prevLocation, change.value),
-            startLocation
-        );
+
+        if (toRef) {
+            // We have a NavRef. We are going to use that.
+            if ((toRef as any).call) {
+                // This is a NavRef object, not a call. We are supposed to execute it now.
+                const navRef = toRef as NavRef<any>;
+                const { location } = navRef.createDirectLink();
+                return location;
+            } else {
+                // This ia a NavRefCall.
+                const { navRef, tokens } = toRef as NavRefCall<any>;
+                const { location } = navRef.createDirectLink(tokens);
+                return location;
+            }
+        } else {
+            // No navRef, so we will use the token-based parameters.
+            const startLocation = appendTokens
+                ? locationManager.getNewLocationForAppendedPathTokens(appendTokens)
+                : removeTokenCount
+                ? locationManager.getNewLocationForRemovedPathTokens(removeTokenCount)
+                : locationManager.getNewLocationForPathAndQueryChanges(undefined, position, to, undefined);
+            return (argChanges || []).reduce(
+                (prevLocation: MyLocation, change) => change.arg.getModifiedLocation(prevLocation, change.value),
+                startLocation
+            );
+        }
     }
 
     protected handleClick() {